% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNADEscore.R
\name{getCNADEscore}
\alias{getCNADEscore}
\title{getCNADEscore}
\usage{
getCNADEscore(inexpData, CNALabel)
}
\arguments{
\item{inexpData}{Gene Expression profile(GEP)}

\item{CNALabel}{The CNA label (the result of function getCNAlabel)}
}
\value{
A matrix of the Student’s t-test result and CNA-DEscore
}
\description{
The function getCNADEscore used calculate the CNA-DEscore.The CNA-DEscore is defined as differentially expressed genes driven by CNA. We used the Student’s t-test to calculate the t scores with the CNA label, and standardized them to a normal distribution with a mean of 0 and an SD of 1.
}
\examples{
# obtain GEP data
GEP<-GetExampleData("GEP")
# obtain the result of getCNAlabel function
CNAlabelresult<-GetExampleData("CNAlabelresult")
# Calculate the CNA-DEscore
CNADEscore<-getCNADEscore(GEP,CNAlabelresult)
# view first the ten gene CNA-DEscore
head(CNADEscore[1:10,])
}
