% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExampleData.R
\name{GetExampleData}
\alias{GetExampleData}
\title{GetExampleData}
\usage{
GetExampleData(exampleData)
}
\arguments{
\item{exampleData}{A character,should be one of"ampGene","delGene","GEP","CNAP", "SubPathwayInfo" ,"GoInfo","GoSubPconGene","Jaccardscore","SubPathwaymapdata","CNAlabelresult","CNADEscore"and "CNA2Subpathwayresult".}
}
\value{
data
}
\description{
Get the example data of test package for litte trials
}
