% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDSubPath.R
\name{IDSubPath}
\alias{IDSubPath}
\title{IDSubPath}
\usage{
IDSubPath(CNADEscore, nperm, Subpathway, Go, Jaccard, Go_SubPath_gene)
}
\arguments{
\item{CNADEscore}{The CNA-DEscore of genes.}

\item{nperm}{input Number of disturbance(>=1000).}

\item{Subpathway}{Subpathway information from SuupathwayMiner.}

\item{Go}{Biological functions data from Gene Ontology.}

\item{Jaccard}{Jaccard score shared by a sub-pathway with biological function. Used to optimization.}

\item{Go_SubPath_gene}{Genes symble shared by a sub-pathway with biological function. Used to optimization.}
}
\value{
Optimized subpathway,and the statistical significance p value and FDR for these optimal subpathways
}
\description{
The function IDSubPath used to find significant dysregulate subpathways by the IDSCNA method.
According our method, in this function, the user need input 6 variable. Four of six variable can obtain from our example data, those data from KEGG and GO, and the user can change at will also. The "CNADEscore" variable is result of function getCNADEscore. And the last variable is nperm, that representative number of disturbances, usually nperm = 1000 or bigger as statistically significant.
}
\examples{

# load depend package
require(igraph)
# obtain subpathway data
SubPathwayInfo<-GetExampleData("SubPathwayInfo")
# obtain biological functions data
GoInfo<-GetExampleData("GoInfo")
# obtain some genes shared by a sub-pathway with biological function
GoSubPconGene<-GetExampleData("GoSubPconGene")
# obtain Jaccardscore shared by a sub-pathway with biological function
Jaccardscore<-GetExampleData("Jaccardscore")
# obtain the result of getDEGscore function
CNADEscore<-GetExampleData("CNADEscore")
# run the IDSubPath function
#CNA2Subpathwayresult<-IDSubPath(CNADEscore,nperm=100,
#                     Subpathway=SubPathwayInfo,Go=GoInfo,
#                     Jaccard=Jaccardscore,Go_SubPath_gene=GoSubPconGene)
# get the result of the IDSCNA function #only show
CNA2Subpathwayresult<-GetExampleData("CNA2Subpathwayresult")
# View the result top 10 subpathways
head(CNA2Subpathwayresult[1:10,])
}
