% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiscovermap.R
\name{plotDiscovermap}
\alias{plotDiscovermap}
\title{plotDiscovermap}
\usage{
plotDiscovermap(genes, CNAmatrix)
}
\arguments{
\item{genes}{A vector of strings. The user intresting and happened CNA genes symbol.}

\item{CNAmatrix}{The copy number alteration matrix. It takes the from of a binary matrix. Only 0 ang 1 are allowed as values.}
}
\value{
Groupwise test of the discover result
}
\description{
we use Copy Number Alteration data to polt discover.
}
\examples{
# load depend package
require(discover)
# input gene
genes <- c("AKT3","AKT1","CALML5","PIK3CA","PIK3R1","CALM1","CALM3","CALML3")
# get the result of the getCNAlabel function
CNAlabelresult<-GetExampleData("CNAlabelresult")
# plot discover
discover<-plotDiscovermap(genes,CNAlabelresult)
plot(discover)
}
