% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNADEscore.R
\name{getCNADEscore}
\alias{getCNADEscore}
\title{getCNADEscore}
\usage{
getCNADEscore(inexpData, CNALabel)
}
\arguments{
\item{inexpData}{Gene Expression profile(GEP)}

\item{CNALabel}{The CNA label (the result of function getCNAlabel)}
}
\value{
A matrix of the Student’s t-test result and CNA-DEscore
}
\description{
The function getCNADEscore used calculate the CNA-DEscore.we use Student’s t-test method to compute the gene differentially expressed level between CNA labels. To compare the genes regulated by CNAs at a common level, we convert the t-test p-value of each gene to z-score.
}
\examples{
# obtain GEP data
GEP<-GetExampleData("GEP")
# obtain the result of getCNAlabel function
CNAlabelresult<-GetExampleData("CNAlabelresult")
# Calculate the CNA-DEscore
CNADEscore<-getCNADEscore(GEP,CNAlabelresult)
# view first the ten gene CNA-DEscore
head(CNADEscore[1:10,])
}
